+++ cpp/tensorrt_llm/runtime/mcastDeviceMemory.h
@@ -17,6 +17,7 @@
 
 #include "tensorrt_llm/common/mcastDevMemUtils.h"
 #include "tensorrt_llm/runtime/ipcNvlsMemory.h"
+#include "tensorrt_llm/runtime/torchUtils.h"
 #include <cstddef>
 #include <cstdint>
 #include <cuda.h>
@@ -106,5 +107,75 @@ private:
 };
 
 constexpr size_t kSIGNAL_PAD_SIZE = 2048;
+
+//! \brief Wrapper class for McastDeviceMemory to facilitate PyTorch tensor creation.
+//! It manages a buffer accessible via unicast or multicast for multi-node communication.
+class McastGPUBuffer
+{
+public:
+    // Disallow copy construction and assignment
+    McastGPUBuffer(McastGPUBuffer const&) = delete;
+    McastGPUBuffer& operator=(McastGPUBuffer const&) = delete;
+
+    //! \brief Constructor for McastGpuBuffer.
+    //! \param bufSize The total size of the buffer in bytes.
+    //! \param groupSize The number of ranks in the communication group.
+    //! \param groupRank The rank of the local process within the group.
+    //! \param device The CUDA device for buffer allocation.
+    //! \param mnNvlink Flag indicating if multi-node NVLink is used.
+    McastGPUBuffer(size_t bufSize, uint32_t groupSize, uint32_t groupRank, at::Device device, bool mnNvlink)
+        : mMcastDeviceMemory(bufSize, groupSize, groupRank, device.index(), mnNvlink)
+        , mGroupSize(groupSize)
+        , mGroupRank(groupRank)
+        , mBufSize(bufSize)
+        , mLocalDevice(device)
+    {
+    }
+
+    //! \brief Returns a PyTorch tensor view of the unicast buffer portion for a specific rank.
+    //! \param rank The target rank for the unicast pointer.
+    //! \param sizes The desired shape (dimensions) of the tensor.
+    //! \param dtype The data type of the tensor elements.
+    //! \param storageOffset The offset in elements from the start of the buffer.
+    //! \return An ATen tensor wrapping the unicast buffer section.
+    at::Tensor getUCBuffer(uint32_t rank, c10::IntArrayRef sizes, c10::ScalarType dtype, int64_t storageOffset)
+    {
+        size_t const numel = std::accumulate(sizes.begin(), sizes.end(), 1UL, std::multiplies<size_t>());
+        size_t const elementSize = c10::elementSize(dtype);
+        size_t const reqSize = (numel + storageOffset) * elementSize;
+        TORCH_CHECK(reqSize <= mBufSize, "McastGpuBuffer::getUcBuffer: the requested size (", reqSize,
+            " bytes) exceeds the allocated size (", mBufSize, " bytes)");
+        auto* dataPtr = static_cast<uint8_t*>(mMcastDeviceMemory.getUnicastPtr(rank)) + storageOffset * elementSize;
+
+        auto options = at::TensorOptions().dtype(dtype).device(mLocalDevice);
+        return at::for_blob(dataPtr, sizes).options(options).target_device(mLocalDevice).make_tensor();
+    }
+
+    //! \brief Returns a PyTorch tensor view of the multicast buffer portion.
+    //! \param sizes The desired shape (dimensions) of the tensor.
+    //! \param dtype The data type of the tensor elements.
+    //! \param storageOffset The offset in elements from the start of the buffer.
+    //! \return An ATen tensor wrapping the multicast buffer section.
+    at::Tensor getMCBuffer(c10::IntArrayRef sizes, c10::ScalarType dtype, int64_t storageOffset)
+    {
+        size_t const numel = std::accumulate(sizes.begin(), sizes.end(), 1UL, std::multiplies<size_t>());
+        size_t const elementSize = c10::elementSize(dtype);
+        size_t const reqSize = (numel + storageOffset) * elementSize;
+        TORCH_CHECK(reqSize <= mBufSize, "McastGpuBuffer::getMcBuffer: the requested size (", reqSize,
+            " bytes) exceeds the allocated size (", mBufSize, " bytes)");
+        auto* dataPtr = static_cast<uint8_t*>(mMcastDeviceMemory.getMulticastPtr()) + storageOffset * elementSize;
+
+        auto options = at::TensorOptions().dtype(dtype).device(mLocalDevice);
+        return at::for_blob(dataPtr, sizes).options(options).target_device(mLocalDevice).make_tensor();
+    }
+
+private:
+    //!< Underlying memory manager for multi-node communication.
+    tensorrt_llm::runtime::McastDeviceMemory mMcastDeviceMemory;
+    uint32_t mGroupSize;     //!< Size of the communication group.
+    uint32_t mGroupRank;     //!< Rank of the current process.
+    size_t mBufSize;         //!< Total size of the managed buffer.
+    at::Device mLocalDevice; //!< The local CUDA device.
+};
 
 } // namespace tensorrt_llm::runtime
