+++ tensorrt_llm/_torch/distributed/__init__.py
@@ -1,15 +1,13 @@
 from tensorrt_llm.functional import AllReduceFusionOp
 
 from .communicator import Distributed, MPIDist, PPComm, TorchDist
-from .ops import (AllReduce, AllReduceParams, AllReduceStrategy,
-                  LowLatencyTwoShotAllReduce, MoEAllReduce, allgather,
-                  reducescatter, userbuffers_allreduce_finalize)
+from .ops import (AllReduce, AllReduceParams, AllReduceStrategy, MoEAllReduce,
+                  allgather, reducescatter, userbuffers_allreduce_finalize)
 
 __all__ = [
     "allgather",
     "reducescatter",
     "userbuffers_allreduce_finalize",
-    "LowLatencyTwoShotAllReduce",
     "AllReduce",
     "AllReduceParams",
     "AllReduceFusionOp",
