+++ tensorrt_llm/_torch/distributed/ops.py
@@ -63,7 +63,7 @@ def get_allreduce_mnnvl_workspace(
     if mapping not in allreduce_mnnvl_workspaces:
         # buffer_tokens * hidden_dim * 3 * 2 * dtype.itemsize
         stride = 3 * 2 * dtype.itemsize
-        buffer_size_in_bytes = math.ceil(12_000_000 / stride) * stride
+        buffer_size_in_bytes = math.ceil(11_010_048 / stride) * stride
         max_num_elements = buffer_size_in_bytes // stride
 
         mcast_buffer = McastGPUBuffer(
@@ -487,127 +487,3 @@ class MoEAllReduce(nn.Module):
             nranks=self.mapping.tp_size,
             eps=eps,
         )
-
-
-class LowLatencyTwoShotAllReduce(nn.Module):
-    # Use singleton pattern since memory allocation is required and expensive
-    # TODO: Do we need to limit the number of instances?
-    _instances = {}
-    _lock = threading.Lock()
-
-    def __new__(cls,
-                mapping: Mapping,
-                init_dim: int = 7168,
-                dtype: torch.dtype = torch.bfloat16):
-        with cls._lock:
-            if init_dim not in cls._instances:
-                cls._instances[init_dim] = super().__new__(cls)
-                cls._instances[init_dim]._initialized = False
-        # __init__ is called after returning this instance
-        return cls._instances[init_dim]
-
-    def __init__(self,
-                 mapping: Mapping,
-                 init_dim: int = 7168,
-                 dtype: torch.dtype = torch.bfloat16):
-        # Prevent re-initialization if __init__ is called again on the same instance
-        if getattr(self, "_initialized", False):
-            return
-        super().__init__()
-
-        self.tp_size = mapping.tp_size
-        self.tp_rank = mapping.tp_rank
-        self.gpus_per_node = mapping.gpus_per_node
-        self.dtype = dtype
-
-        # Use intra-node to avoid confusing and avoid using 'Local Rank'
-        self.force_mn = os.environ.get("TRTLLM_LLAR_FORCE_MN", "0") == "1"
-        self.intra_node_rank = self.tp_rank % self.gpus_per_node
-        self.local_device = torch.device("cuda", self.intra_node_rank)
-        self.is_multi_node = mapping.is_multi_node() or self.force_mn
-        self.max_num_tokens = int(os.environ.get("TRTLLM_LLAR_MAX_M", "128"))
-
-        # Predefined N used to align the allocation
-        self.hidden_dim = init_dim
-        self.buf_size = self._alloc_buf()
-
-        self._initialized = True
-
-    def _alloc_buf(self):
-        # Triple-buffer, one buffer for the reduce-scatter and one for the allgather, M*N
-        buffer_tokens = math.ceil(
-            self.max_num_tokens / self.tp_size) * self.tp_size
-        self._mcast_buffer = McastGPUBuffer(
-            buffer_tokens * self.hidden_dim * 3 * 2 * self.dtype.itemsize,
-            self.tp_size,
-            self.tp_rank,
-            self.local_device,
-            self.is_multi_node,
-        )
-
-        self._buffer = self._mcast_buffer.get_uc_buffer(
-            self.tp_rank, (3, 2, buffer_tokens, self.hidden_dim), self.dtype, 0)
-        # Only initialize the buffer when we need to resize it
-        self._buffer.fill_(-0.0)
-        # CPU barrier since we assume this should not be called in cuda graph
-        torch.cuda.synchronize()
-        mpi_barrier()
-
-        # This is a buffer to maintain the state of this allreduce Op
-        # Should have the same lifetime with self._buffer
-        # [Buffer_ptr, Clear_ptr, Buffer_size, atomic access counter]
-        self._buffer_flags = torch.tensor(
-            [0, 2, self._buffer.size()[3] * self._buffer.size()[2], 0],
-            dtype=torch.uint32,
-            device=self.local_device)
-
-        return buffer_tokens * self.hidden_dim
-
-    def __call__(self, shard_in: torch.Tensor) -> torch.Tensor:
-        buffer_stride = self._buffer.size()[3] * self._buffer.size(
-        )[2] * self._buffer.size()[1]
-        shape = shard_in.shape
-        shard_in = shard_in.view(-1, shard_in.shape[-1])
-
-        shard_out = torch.empty_like(shard_in)
-
-        torch.ops.trtllm.lowlat_twoshot_allreduce(
-            shard_out,
-            shard_in,
-            self._buffer,
-            self._buffer_flags,
-            True,
-        )
-        return shard_out.view(shape)
-
-    def all_reduce_res_norm(
-        self,
-        gamma: torch.Tensor,
-        x: torch.Tensor,
-        residual_in: torch.Tensor,
-        residual_out: Optional[torch.Tensor] = None,
-        eps: float = torch.finfo(torch.bfloat16).eps,
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        buffer_stride = self._buffer.size()[3] * self._buffer.size(
-        )[2] * self._buffer.size()[1]
-        shape = x.shape
-        x_flattened = x.view(-1, x.shape[-1])
-        shard_out = torch.empty_like(x_flattened)
-        residual_in = residual_in.view(-1, residual_in.shape[-1])
-        assert x_flattened.shape[-1] == self.hidden_dim
-
-        torch.ops.trtllm.lowlat_twoshot_allreduce(
-            shard_out,
-            x_flattened,
-            self._buffer,
-            self._buffer_flags,
-            False,
-        )
-        if residual_out is None:
-            residual_out = torch.empty_like(residual_in)
-
-        torch.ops.trtllm.lowlat_twoshot_rmsnorm(residual_out, shard_out,
-                                                self._buffer, gamma, eps,
-                                                residual_in, self._buffer_flags)
-
-        return shard_out.view(shape), residual_out.view(shape)
